requirements.txt
aiohttp
pandas
tqdm
Pillow
numpy